+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
-**./NameSorter.Tests/ExampleTest.cs**
-* Deleted file
+**./NameSorter/Common/NameOrder.cs**
+* Sorting can no longer be performed by original line order.
+	* Removed 'orderComp' variable.
+	* If the names are the same, 'resultFlag' is 0
+	* Changed "Sort by original line order." to "Tied"